use macros to use faster versions
